pyproject.toml
[tool.poetry]
name = "super-resolution-bot"
version = "0.1.0"
description = ""
authors = ["Gamazic <nikita.nov.ru@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.19"
celery = "^5.2.3"
redis = "^4.1.3"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
